+++ backend/src/main/java/com/overlang/api/dto/auth/AuthResponse.java
@@ -1,0 +1,8 @@
+package com.overlang.api.dto.auth;
+
+import io.swagger.v3.oas.annotations.media.Schema;
+
+@Schema(description = "로그인 인증 응답 DTO")
+public record AuthResponse(
+    @Schema(description = "액세스 토큰") String accessToken,
+    @Schema(description = "리프레시 토큰") String refreshToken) {}
