+++ backend/src/main/java/com/overlang/domain/common/BaseTimeEntity.java
@@ -1,0 +1,22 @@
+package com.overlang.domain.common;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.EntityListeners;
+import jakarta.persistence.MappedSuperclass;
+import java.time.Instant;
+import lombok.Getter;
+import org.springframework.data.annotation.CreatedDate;
+import org.springframework.data.annotation.LastModifiedDate;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+
+@Getter
+@MappedSuperclass
+@EntityListeners(AuditingEntityListener.class)
+public abstract class BaseTimeEntity {
+
+  @CreatedDate
+  @Column(updatable = false)
+  private Instant createdAt; // ISO 8601 / UTC
+
+  @LastModifiedDate private Instant updatedAt; // ISO 8601 / UTC
+}
