+++ backend/src/main/java/com/overlang/domain/job/entity/JobStatus.java
@@ -1,0 +1,8 @@
+package com.overlang.domain.job.entity;
+
+public enum JobStatus {
+  PENDING, // 대기 중
+  RUNNING, // 처리 중
+  COMPLETED, // 완료
+  FAILED // 실패
+}
