+++ backend/src/main/java/com/overlang/domain/member/repository/MemberRepository.java
@@ -1,0 +1,8 @@
+package com.overlang.domain.member.repository;
+
+import com.overlang.domain.member.entity.Member;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface MemberRepository extends JpaRepository<Member, Long> {}
