+++ backend/src/main/java/com/overlang/global/config/WebConfig.java
@@ -1,0 +1,20 @@
+package com.overlang.global.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.CorsRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Configuration
+public class WebConfig implements WebMvcConfigurer {
+
+  @Override
+  public void addCorsMappings(CorsRegistry registry) {
+    registry
+        .addMapping("/**") // 모든 API 경로에 대해
+        .allowedOrigins("http://localhost:3000") // 프론트엔드 개발 서버 주소
+        .allowedMethods("GET", "POST", "PUT", "DELETE", "PATCH", "OPTIONS")
+        .allowedHeaders("*")
+        .allowCredentials(true)
+        .maxAge(3600); // 브라우저가 이 설정을 1시간 동안 기억
+  }
+}
