+++ backend/src/main/java/com/overlang/global/response/ErrorResponse.java
@@ -1,0 +1,20 @@
+package com.overlang.global.response;
+
+public class ErrorResponse {
+
+  private String code;
+  private String message;
+
+  public ErrorResponse(String code, String message) {
+    this.code = code;
+    this.message = message;
+  }
+
+  public String getCode() {
+    return code;
+  }
+
+  public String getMessage() {
+    return message;
+  }
+}
