--- backend/src/main/java/kr/ac/hansung/cse/overlang/OverlangBackendApplication.java
@@ -1,13 +1,0 @@
-package kr.ac.hansung.cse.overlang;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@EnableJpaAuditing
-@SpringBootApplication
-public class OverlangBackendApplication {
-  public static void main(String[] args) {
-    SpringApplication.run(OverlangBackendApplication.class, args);
-  }
-}
