--- backend/src/main/java/kr/ac/hansung/cse/overlang/domain/job/entity/JobStatus.java
@@ -1,8 +1,0 @@
-package kr.ac.hansung.cse.overlang.domain.job.entity;
-
-public enum JobStatus {
-  PENDING, // 대기 중
-  RUNNING, // 처리 중
-  COMPLETED, // 완료
-  FAILED // 실패
-}
