--- backend/src/main/java/kr/ac/hansung/cse/overlang/domain/member/repository/MemberRepository.java
@@ -1,8 +1,0 @@
-package kr.ac.hansung.cse.overlang.domain.member.repository;
-
-import kr.ac.hansung.cse.overlang.domain.member.entity.Member;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface MemberRepository extends JpaRepository<Member, Long> {}
