--- backend/src/main/java/kr/ac/hansung/cse/overlang/global/advice/GlobalExceptionHandler.java
@@ -1,17 +1,0 @@
-package kr.ac.hansung.cse.overlang.global.advice;
-
-import kr.ac.hansung.cse.overlang.global.response.ApiResponse;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-@RestControllerAdvice
-public class GlobalExceptionHandler {
-
-  @ExceptionHandler(Exception.class)
-  public ResponseEntity<ApiResponse<Void>> handleException(Exception e) {
-    return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
-        .body(ApiResponse.error("COMMON_001", e.getMessage()));
-  }
-}
