--- backend/src/main/java/kr/ac/hansung/cse/overlang/global/response/ErrorResponse.java
@@ -1,20 +1,0 @@
-package kr.ac.hansung.cse.overlang.global.response;
-
-public class ErrorResponse {
-
-  private String code;
-  private String message;
-
-  public ErrorResponse(String code, String message) {
-    this.code = code;
-    this.message = message;
-  }
-
-  public String getCode() {
-    return code;
-  }
-
-  public String getMessage() {
-    return message;
-  }
-}
