+++ backend/src/test/java/com/overlang/OverlangBackendApplicationTests.java
@@ -1,0 +1,11 @@
+package com.overlang;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class OverlangBackendApplicationTests {
+
+  @Test
+  void contextLoads() {}
+}
