--- backend/src/test/java/kr/ac/hansung/cse/overlang/OverlangBackendApplicationTests.java
@@ -1,11 +1,0 @@
-package kr.ac.hansung.cse.overlang;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class OverlangBackendApplicationTests {
-
-  @Test
-  void contextLoads() {}
-}
